Allow system-test to stop on error
Added a env flag CONTIV_SOE that can be specified with the
make system-test and regress-test targets to stop the test on first
failure and leave the system in same state for debugging. This flag
shall be useful for debugging test issues seen during development
and local testing.

+++ systemtests/utils/docker.go
@@ -20,7 +20,10 @@ import (
 	"testing"
 )
 
-func DockerCleanup(node VagrantNode, contName string) {
+func DockerCleanup(t *testing.T, node VagrantNode, contName string) {
+	if !OkToCleanup(t.Failed()) {
+		return
+	}
 	cmdStr := "sudo docker kill " + contName
 	node.RunCommand(cmdStr)
 	cmdStr = "sudo docker rm " + contName
